Initial UE5.4.2 Commit

+++ Source/GraphicSettings/Private/GraphicSettingsSubsystem.cpp
@@ -6,42 +6,42 @@
 
 void UGraphicSettingsSubsystem::SetAntiAliasingQuality(EGraphicSettingsQuality Quality)
 {
-	GEngine->GameUserSettings->SetAntiAliasingQuality(Quality);
+	GEngine->GameUserSettings->SetAntiAliasingQuality((int32)Quality);
 
 	GEngine->GameUserSettings->ApplySettings(true);
 }
 
 void UGraphicSettingsSubsystem::SetFoliageQuality(EGraphicSettingsQuality Quality)
 {
-	GEngine->GameUserSettings->SetFoliageQuality(Quality);
+	GEngine->GameUserSettings->SetFoliageQuality((int32)Quality);
 
 	GEngine->GameUserSettings->ApplySettings(true);
 }
 
 void UGraphicSettingsSubsystem::SetPostProcessingQuality(EGraphicSettingsQuality Quality)
 {
-	GEngine->GameUserSettings->SetPostProcessingQuality(Quality);
+	GEngine->GameUserSettings->SetPostProcessingQuality((int32)Quality);
 
 	GEngine->GameUserSettings->ApplySettings(true);
 }
 
 void UGraphicSettingsSubsystem::SetShadowQuality(EGraphicSettingsQuality Quality)
 {
-	GEngine->GameUserSettings->SetShadowQuality(Quality);
+	GEngine->GameUserSettings->SetShadowQuality((int32)Quality);
 
 	GEngine->GameUserSettings->ApplySettings(true);
 }
 
 void UGraphicSettingsSubsystem::SetTextureQuality(EGraphicSettingsQuality Quality)
 {
-	GEngine->GameUserSettings->SetTextureQuality(Quality);
+	GEngine->GameUserSettings->SetTextureQuality((int32)Quality);
 
 	GEngine->GameUserSettings->ApplySettings(true);
 }
 
 void UGraphicSettingsSubsystem::SetVisualEffectQuality(EGraphicSettingsQuality Quality)
 {
-	GEngine->GameUserSettings->SetVisualEffectQuality(Quality);
+	GEngine->GameUserSettings->SetVisualEffectQuality((int32)Quality);
 
 	GEngine->GameUserSettings->ApplySettings(true);
 }
